phase1_verification/pyproject.toml
[project]
name = "taxlaw-kg-phase1"
version = "0.1.0"
description = "Phase 1 verification for Tax Law Knowledge Base"
requires-python = ">=3.10"
dependencies = [
    "lightrag-hku>=1.0.0",
    "dashscope>=1.20.0",
    "neo4j>=5.0.0",
    "redis>=5.0.0",
    "pymilvus>=2.4.0",
    "numpy>=1.26.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "aiofiles>=23.0.0",
    "httpx>=0.26.0",
]

[tool.uv]
dev-dependencies = []
